move blog domain class to model package
