+++ app/src/main/java/com/droid/diexample/data/BlogRepository.kt
@@ -1,13 +1,9 @@
 package com.droid.diexample.data
 
-import com.droid.diexample.domain.BlogDomain
+import com.droid.diexample.domain.model.BlogDomain
 import com.droid.diexample.data.remote.api.BlogApiService
-import com.droid.diexample.data.local.dao.BlogDao
-import com.droid.diexample.data.local.mapper.CacheMapper
 import com.droid.diexample.data.remote.mappers.toDomain
 import com.droid.diexample.domain.repository.IBlogsRepository
-import com.droid.diexample.util.DataState
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
