+++ app/src/main/java/com/droid/diexample/data/remote/mappers/JsonResponseToDomain.kt
@@ -1,9 +1,9 @@
 package com.droid.diexample.data.remote.mappers
 
 import com.droid.diexample.data.remote.model.Blog
-import com.droid.diexample.domain.BlogDomain
+import com.droid.diexample.domain.model.BlogDomain
 
-internal fun Blog.toDomain():BlogDomain =
+internal fun Blog.toDomain(): BlogDomain =
     BlogDomain(
         this.id,
         this.title,
